MAT-8639: Move test case validation logic into own service. Co-locating the test case validation, async, and concurrency logic prevents circular dependencies.
